Moved repairTag method and ran code formatter
